+++ src/server/config/heatingMode.ts
@@ -1,0 +1,5 @@
+export enum HeatingMode {
+  None,
+  Winter,
+  Sommer,
+}