+++ src/server/config/iAsusConfig.ts
@@ -1,0 +1,6 @@
+export interface iAsusConfig {
+  address: string;
+  ignoreSSL?: boolean;
+  password: string;
+  username: string;
+}