+++ src/server/config/iBlueIrisSettings.ts
@@ -1,0 +1,5 @@
+export interface iBlueIrisSettings {
+  serverAddress: string;
+  username?: string;
+  password?: string;
+}