+++ src/server/config/iBluetoothTrackingSettings.ts
@@ -1,0 +1,4 @@
+export interface iBluetoothTrackingSettings {
+  customName: string;
+  activeTracking: boolean;
+}