+++ src/server/config/iDaikinSettings.ts
@@ -1,0 +1,7 @@
+export interface iDaikinSettings {
+  active: boolean;
+  activateTracingLogger?: boolean;
+  useGetToPost?: boolean;
+  // Whether Room Default Callbacks should be added to start ac on Bottom Right short and stop on long button press.
+  buttonBotRightForAc?: boolean;
+}