+++ src/server/config/iEspresenseSettings.ts
@@ -1,0 +1,6 @@
+import { iBluetoothTrackingSettings } from './iBluetoothTrackingSettings';
+
+export interface iEspresenseSettings {
+  mqttInstance: number;
+  deviceMap: { [folderId: string]: iBluetoothTrackingSettings };
+}