+++ src/server/config/iMp3Settings.ts
@@ -1,0 +1,6 @@
+export interface iMp3Settings {
+  // local path for the mp3 files to store/load
+  path: string;
+  // external reachable adress to access those mp3 files
+  serverAddress: string;
+}