+++ src/server/config/iPersistenceSettings.ts
@@ -1,0 +1,5 @@
+import { PoolConfig } from 'pg';
+
+export interface iPersistenceSettings {
+  postgreSql?: PoolConfig;
+}