+++ src/server/config/iPollySettings.ts
@@ -1,0 +1,8 @@
+export interface iPollySettings {
+  mp3Path: string;
+  region: string;
+  signatureVersion: string;
+  accessKeyId: string;
+  secretAccessKey: string;
+  voiceID: string;
+}