+++ src/server/config/iRestSettings.ts
@@ -1,0 +1,4 @@
+export interface iRestSettings {
+  active: boolean;
+  port?: number;
+}