+++ src/server/config/iSonosSettings.ts
@@ -1,0 +1,5 @@
+export interface iSonosSettings {
+  active: boolean;
+  // Whether Room Default Callbacks should be added to start Radio on Bottom Right long button press.
+  buttonBotRightForRadio?: boolean;
+}