+++ src/server/config/iTelegramSettings.ts
@@ -1,0 +1,8 @@
+import { LogLevel } from '../../models';
+
+export interface iTelegramSettings {
+  logLevel: LogLevel;
+  telegramToken: string;
+  allowedIDs: number[];
+  subscribedIDs: number[];
+}