+++ src/server/config/iTimePair.ts
@@ -1,0 +1,4 @@
+export interface iTimePair {
+  hours: number;
+  minutes: number;
+}