+++ src/server/config/iTimeSettings.ts
@@ -1,0 +1,6 @@
+import { iTimePair } from './iTimePair';
+
+export interface iTimeSettings {
+  nightStart: iTimePair;
+  nightEnd: iTimePair;
+}