+++ src/server/config/iWeatherSettings.ts
@@ -1,0 +1,6 @@
+export interface iWeatherSettings {
+  lattitude: string;
+  longitude: string;
+  // App-ID from OpenWeatherMap
+  appid?: string;
+}